Moved common functions to utilities module
