+++ makeAllEditions.py
@@ -1,7 +1,4 @@
-def collect_content_directory_names():
-    import os
-    content_directory_prefix = "edition_"
-    return [name for name in os.listdir(".") if os.path.isdir(name) and name.startswith(content_directory_prefix)]
+from utilities import collect_content_directory_names
 
 if __name__ == "__main__":
     from makeNewsletter import make_newsletter
