+++ makeNewsletter.py
@@ -3,19 +3,9 @@ from markdown.extensions.toc import TocExtension
 import os, shutil
 import re
 import argparse
+from utilities import enforce_default_encoding, get_source_image_directory_path, get_destination_image_directory_path
 
-import sys
-# in case the system default encoding is not utf-8: it may cause UnicodeEncodeError on Markdown.convert() or file.write()
-if sys.getdefaultencoding() is not "utf-8":
-    try:
-        print("change encoding from {} to utf-8".format(sys.getdefaultencoding()))
-        sys.setdefaultencoding("utf-8")
-    except AttributeError:
-        reload(sys) # I don't know why but this fixes "AttributeError: 'module' object has no attribute 'setdefaultencoding'"
-        sys.setdefaultencoding("utf-8")
-    except Exception as e:
-        print("changing encoding failed: {}".format(e))
-
+enforce_default_encoding()
 
 def make_newsletter(target_edition_name):
     if not os.path.isdir(target_edition_name):
@@ -66,11 +56,9 @@ def make_newsletter(target_edition_name):
         o.write(htmltext)
 
     # copying images
-    source_image_dir = os.path.join(target_edition_name, "images")
-    dest_image_dir = os.path.join("docs", "images_"+target_edition_name)
+    source_image_dir = get_source_image_directory_path(target_edition_name)
+    dest_image_dir = get_destination_image_directory_path(target_edition_name)
 
-    if not os.path.isdir(dest_image_dir):
-        os.mkdir(dest_image_dir)
     for file in os.listdir(source_image_dir):
         shutil.copyfile(os.path.join(source_image_dir, file),
                         os.path.join(dest_image_dir, file))
